backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0
python-multipart==0.0.6
pydantic[email]==2.5.3
aiofiles==23.2.1
python-dotenv==1.0.0
